requirements.txt
google_api_python_client==2.147.0
google_auth_oauthlib==1.2.1
protobuf==5.28.2
PyQt5==5.15.11
PyQt5_sip==12.15.0
cryptography==43.0.1